% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_bb}
\alias{as_bb}
\title{Retrieves the extent of an object's bonding box}
\usage{
as_bb(x)
}
\arguments{
\item{x}{numeric, SpatVector, SpatRaster, sf or stars object}
}
\value{
numeric vector of bounding box in [xmin, xmax, ymin, ymax] order
}
\description{
Retrieves the extent of an object's bonding box
}
