% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo1.R
\name{etopo_nc_nav}
\alias{etopo_nc_nav}
\title{Retrieve GEBCO ncdf navigation values (start, count, lons, lats)}
\usage{
etopo_nc_nav(x, bb = c(-72, -63, 39, 46), res = etopo_nc_res(), varname = "z")
}
\arguments{
\item{x}{ncdfd4 object}

\item{bb}{numeric, 4 element requested bounding box [west, east, south, north]}

\item{res}{numeric, 2 element resolution [res_x,res_y]}

\item{varname}{character the name of the variable}
}
\value{
list with
\itemize{
  \item{bb the requested bounding box}
  \item{res the resolution}
  \item{start vector of start values}
  \item{count vector of count values}
  \item{ext vector of extent }
  \item{crs character}
  \item{varname character}
}
}
\description{
Retrieve GEBCO ncdf navigation values (start, count, lons, lats)
}
