% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo1.R
\name{etopo_path}
\alias{etopo_path}
\title{Retrieve the etopo path}
\usage{
etopo_path(..., root = topo_path("etopo"))
}
\arguments{
\item{...}{file path segments to append to the root}

\item{root}{character, the root directory}
}
\value{
file path specification (which may not exists)
}
\description{
Retrieve the etopo path
}
