% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopotools-package.R
\docType{package}
\name{etopotools}
\alias{etopotools}
\alias{etopo}
\alias{etopo1}
\title{etopotools - R tools for use with right whale observations compiled by Hansen Johnson and others.}
\description{
Coding tools for working with local copies of \href{https://ngdc.noaa.gov/mgg/global/}{ETOPO data}.
}
