% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gebco.R
\name{gebco_nc_res}
\alias{gebco_nc_res}
\title{Retrieve the stated resolution of the grid}
\usage{
gebco_nc_res(x, lookup = FALSE)
}
\arguments{
\item{x}{ncdfd4 object}

\item{lookup}{logical, if TRUE then retrieve res from the ncdf4 object, otherwise
return a predetermined resolution}
}
\value{
two element numeric vector or [lon, lat] resolution
}
\description{
Retrieve the stated resolution of the grid
}
