% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gebco.R
\name{gebco_path}
\alias{gebco_path}
\title{Retrieve the gebco path}
\usage{
gebco_path(..., root = topo_path("gebco"))
}
\arguments{
\item{...}{file path segments to append to the root}

\item{root}{character, the root directory}
}
\value{
file path sepcification (whihc may not exists)
}
\description{
Retrieve the gebco path
}
