% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopotools-package.R
\name{mask_etopo}
\alias{mask_etopo}
\title{Mask a raster of ETOPO data}
\usage{
mask_etopo(x = read_etopo(), value = 0, where = c("above", "at", "below")[1])
}
\arguments{
\item{x}{raster, as per \code{read_etopo}}

\item{value}{numeric, the value around which to threshold}

\item{where}{character, indiates what gets masked relative to 0
\itemize{
\item{above, values at or above \code{value} are assigned NA, all others assigned 1 - this masks area at or above sealevel}
\item{at, values at \code{value} are assigned NA, above +1 and below -1}
\item{below, values at or below \code{value} are assigned NA, all others assigned 1 - this masks area at or below sealevel}
}}
}
\description{
Mask a raster of ETOPO data
}
