% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topotools-package.R
\name{mask_topo}
\alias{mask_topo}
\title{Mask a raster of topo data}
\usage{
mask_topo(x = read_etopo(), value = 0, where = c("above", "at", "below")[1])
}
\arguments{
\item{x}{raster, as per \code{read_etopo} or \code{read_gebco}}

\item{value}{numeric, the value around which to threshold}

\item{where}{character, indicates what gets masked relative to 0
\itemize{
\item{above, values at or above \code{value} are assigned NA, all others assigned 1 - this masks area at or above sealevel}
\item{at, values at \code{value} are assigned NA, above +1 and below -1}
\item{below, values at or below \code{value} are assigned NA, all others assigned 1 - this masks area at or below sealevel}
}}
}
\value{
RasterLayer or SpatRaster
}
\description{
Mask a raster of topo data
}
