% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopotools-package.R
\name{read_etopo}
\alias{read_etopo}
\title{Read a raster file - possibly subsetting}
\usage{
read_etopo(
  filename = c("ETOPO1_Ice_c_geotiff.tif", "ETOPO1_Ice_c_mask.tif")[1],
  bb = NULL,
  path = etopo_path()
)
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{bb}{numeric (or NULL), 4 element subsetting bounding box [west, east, south, north]}

\item{path}{character, the path to the etopo datasets}
}
\value{
NULL or RasterLayer
}
\description{
Read a raster file - possibly subsetting
}
