% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gebco.R
\name{read_gebco}
\alias{read_gebco}
\title{Read a raster file - possibly subsetting}
\usage{
read_gebco(
  filename = "GEBCO_2022.nc",
  bb = c(-72, -63, 39, 46),
  path = gebco_path()
)
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{bb}{numeric (or NULL), 4 element subsetting bounding box [west, east, south, north]}

\item{path}{character, the path to the etopo datasets}

\item{form}{char, one of 'SpatRaster' or 'stars' (default)}
}
\value{
SpatRaster or stars object
}
\description{
Read a raster file - possibly subsetting
}
