% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resize}
\alias{resize}
\title{Resize a stars raster by a multiplier}
\usage{
resize(x = read_etopo(), m = 2, ...)
}
\arguments{
\item{x}{stars object}

\item{m}{num, a one or two element multiplier to modify the size of cells.
m < 1 increases the number of cells, and m > 1 decreases the number of cells}

\item{...}{other arguments for [stars::st_warp]}
}
\value{
a resized stars object
}
\description{
A warning may be generated by `GDAL` about missing values.  THis warning may be 
ignored in most cases.
}
