% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topotools-package.R
\name{topo_path}
\alias{topo_path}
\title{Retrieve the topo path}
\usage{
topo_path(..., root = "/mnt/ecocast/coredata/bathy")
}
\arguments{
\item{...}{file path segments to append to the root}

\item{root}{character, the root directory}
}
\value{
file path specification (which may not exists)
}
\description{
Retrieve the topo path
}
