% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo1.R
\name{list_etopo}
\alias{list_etopo}
\title{List available files for ETOPO}
\usage{
list_etopo(path = etopo_path(), pattern = "^ETOPO_.*\\\\.tif$")
}
\arguments{
\item{path}{char, the path to the data}

\item{pattern}{char, regular expression of pattern for search}
}
\value{
char fully qualified file paths
}
\description{
List available files for ETOPO
}
