% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gebco.R
\name{list_gebco}
\alias{list_gebco}
\title{List available files for GEBCO}
\usage{
list_gebco(path = gebco_path(), pattern = "^GEBCO_.*\\\\.nc$")
}
\arguments{
\item{path}{char, the path to the data}

\item{pattern}{char, regular expression of pattern for search}
}
\value{
char fully qualified file paths
}
\description{
List available files for GEBCO
}
