% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo1.R
\name{read_etopo}
\alias{read_etopo}
\title{Read a raster file - possibly subsetting}
\usage{
read_etopo(
  filename = c("ETOPO_2022_v1_60s_N90W180_surface.tif",
    "ETOPO_2022_v1_30s_N90W180_surface.tif")[1],
  bb = NULL,
  path = etopo_path(),
  form = c("SpatRaster", "stars")[2]
)
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{bb}{numeric (or NULL), 4 element subsetting bounding box [west, east, south, north]}

\item{path}{character, the path to the etopo datasets}

\item{form}{char, one of 'SpatRaster' or 'stars' (default)}
}
\value{
SpatRaster or stars object
}
\description{
Read a raster file - possibly subsetting
}
